Change PartitionKey to be u32
This commit changes the PartitionKey to be a u32. This reduces
the required bytes for the partition key while still keeping
enough values to scale arbitrarily.

This fixes #279.

+++ src/common/src/partitioner.rs
@@ -8,6 +8,7 @@ impl HashPartitioner {
     pub fn compute_partition_key(value: &impl Hash) -> PartitionKey {
         let mut hasher = xxhash_rust::xxh3::Xxh3::default();
         value.hash(&mut hasher);
-        hasher.finish()
+        // Safe to only take the lower 32 bits: See https://github.com/Cyan4973/xxHash/issues/453#issuecomment-696838445
+        hasher.finish() as u32
     }
 }
