+++ src/common/src/types.rs
@@ -31,7 +31,7 @@ pub type EntryIndex = u32;
 
 /// Identifying to which partition a key belongs. This is unlike the [`PartitionId`]
 /// which identifies a consecutive range of partition keys.
-pub type PartitionKey = u64;
+pub type PartitionKey = u32;
 
 /// Discriminator for invocation instances
 pub type InvocationId = Uuid;
